[refactor/#161] 코드 옮김 + 수정
