+++ src/main/java/com/clokey/server/domain/history/converter/HistoryConverter.java
@@ -146,13 +146,6 @@ public class HistoryConverter {
                 .build();
     }
 
-    public static HistoryResponseDTO.LastYearHistoryResult toLastYearHistoryResult(Long historyId, List<String> historyImageUrls, Member member) {
-        return HistoryResponseDTO.LastYearHistoryResult.builder()
-                .historyId(historyId)
-                .nickName(member.getNickname())
-                .imageUrls(historyImageUrls)
-                .build();
-    }
     public static HistoryResponseDTO.LikedUserResults toLikedUserResult(List<Member> members, List<Boolean> followStatus){
         List<HistoryResponseDTO.LikedUserResult> likedUserResults = new ArrayList<>();
         for(int i=0; i<members.size(); i++){
