+++ src/main/java/com/clokey/server/domain/recommendation/dto/RecommendationResponseDTO.java
@@ -107,5 +107,6 @@ public class RecommendationResponseDTO {
         private String clokeyId;
         private String profileImage;
         private String imageUrl;
+        private Long historyId;
     }
 }
